backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-auth-oauthlib
google-api-python-client
python-dotenv
requests
